Minor moving of files, and bugfixing
