--- backend/src/main/java/no/ntnu/AuthApplication.java
@@ -1,16 +1,0 @@
-package no.ntnu;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-/**
- * The main entrypoint for the application.
- */
-@SpringBootApplication
-public class AuthApplication {
-
-  public static void main(String[] args) {
-    SpringApplication.run(AuthApplication.class, args);
-  }
-
-}
