+++ backend/src/main/java/no/ntnu/database/controllers/CourseController.java
@@ -26,7 +26,6 @@ import org.springframework.web.bind.annotation.RestController;
 @CrossOrigin
 @RestController
 @RequestMapping("/courses")
-
 public class CourseController {
 	private static final Logger LOGGER = LoggerFactory.getLogger(CourseController.class);
