+++ backend/src/main/java/no/ntnu/database/controllers/UserController.java
@@ -1,0 +1,12 @@
+package no.ntnu.database.controllers;
+
+import org.springframework.web.bind.annotation.CrossOrigin;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@CrossOrigin
+@RestController
+@RequestMapping("/users")
+public class UserController {
+
+}
