+++ backend/src/main/java/no/ntnu/database/jwt/AccessUserService.java
@@ -1,10 +1,9 @@
-package no.ntnu.jwt;
+package no.ntnu.database.jwt;
 
 
 import java.util.Optional;
 import no.ntnu.database.entities.User;
 import no.ntnu.database.repositories.UserRepository;
-import no.ntnu.jwt.AccessUserDetails;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
