+++ backend/src/main/java/no/ntnu/database/jwt/SecurityConfiguration.java
@@ -1,7 +1,6 @@
-package no.ntnu.jwt;
+package no.ntnu.database.jwt;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Configurable;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.authentication.AuthenticationManager;
